+++ app/src/main/java/com/fredhappyface/ewesticker/MainActivity.kt
@@ -16,6 +16,7 @@ import android.widget.TextView
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.AppCompatActivity
 import androidx.preference.PreferenceManager
+import com.fredhappyface.ewesticker.utilities.Toaster
 import java.util.*
 import java.util.concurrent.Executors