+++ app/src/main/java/com/fredhappyface/ewesticker/StickerSender.kt
@@ -12,6 +12,9 @@ import androidx.core.view.inputmethod.InputConnectionCompat
 import androidx.core.view.inputmethod.InputContentInfoCompat
 import coil.ImageLoader
 import coil.request.ImageRequest
+import com.fredhappyface.ewesticker.utilities.Cache
+import com.fredhappyface.ewesticker.utilities.Toaster
+import com.fredhappyface.ewesticker.utilities.Utils
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch