+++ app/src/main/java/com/fredhappyface/ewesticker/model/StickerPack.kt
@@ -1,4 +1,4 @@
-package com.fredhappyface.ewesticker
+package com.fredhappyface.ewesticker.model
 
 import java.io.File